move stats to own module for reuse
